# faster project configuration