CPM currently clones the github repo leading to increased download times compared to downloading the zip files:

```
time cmake --preset develop # with cpmaddpackage --> GITHUB_REPOSITORY
real    2m0,627s
user    0m16,555s
sys     0m6,525s

time cmake --preset develop   # with cpmaddpackage --> URL

real    0m30,557s
user    0m5,519s
sys     0m3,345s
```

First introduces by 

@Fitanium 

in PR #163